==> 02_add_movie_specific_info.ipynb <==
### Add in specific features (genre, mpaa rating, runtime, and budget) from each of the 4341 movies
This program uses headless browsing to pull relevant feature information from each movie's main page on BoxOfficeMojo.

import os
import re
import time
import numpy as np
import pandas as pd
import requests
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from bs4 import BeautifulSoup as bs4

# Set-up headless browsing
chrome_options = Options()
chrome_options.add_argument('--disable-extensions')
chrome_options.add_argument('headless')
driver = webdriver.Chrome(executable_path='C:/bin/chromedriver.exe',options=chrome_options)

def add_features_to_list(feature_list,row_list):
    newlist=[]
    for f in feature_list:
        if f in row_list:
            f_index = row_list.index(f)
            newlist.append(row_list[f_index+1])
            continue
        else:
            newlist.append(np.nan)
            continue
    return newlist

features_list=[]

for year in range(1977,2020):
    # Read in each movie's URL from the previously scraped list of movies by year:
    df = pd.read_csv(str(year)+'_domestic_releases.csv')
    
    features=[]
    features_list=[]
    for i in range(0,len(df)):
        if i%20 == 0:
            print(year, i)
        moviepage=str(df.iloc[i][7])
        url='https://www.'+moviepage
        ##print(url)
        driver.get(url)
        html = driver.page_source
        soup = bs4(html, 'html.parser')

        list_text_rows=[]
        
        # Drill down to table area of relevance for capturing features:
        divs = soup.find_all('div',{'class':'mojo-summary-values'})
        for div in divs:
            rows = div.find_all('span')
            for row in rows:
                if(row.text.find('Domestic')):
                    list_text_rows.append(row.text)

        features = ['Budget','Genres','MPAA','Running Time']
        features_list.append(add_features_to_list(features, list_text_rows))
        
        #=================================
        #wait 1 seconds between iterations
        time.sleep(1)

    # Create feature columns and populate:
    temp = list(zip(*features_list))

    # Assign each column the feature values accumulated over all years:
    for i in range(0,3):
        df[features[i]] = list(temp[i])
    
    try:
        # Clean up Genres column
        df['Genres'] = df['Genres'].fillna('-')
        a = [g.rstrip() for g in df['Genres']]
        b = [re.sub('\n\s+',' ',s) for s in a]
        df['Genres'] = b
        
        
        # Write new dataframe to .csv file:
        df.to_csv(str(year)+'_movies_features_TEST.csv', index = False)
        
    except: # Skip a year if error encountered, re-run again separately if needed
        pass

def add_features_to_list(feature_list,row_list):
    newlist=[]
    for f in feature_list:
        if f in row_list:
            f_index = row_list.index(f)
            newlist.append(row_list[f_index+1])
            continue
        else:
            newlist.append(np.nan)
            continue
    return newlist

features_list=[]

# Previously ran for 1977 --> 2010
for year in range(2010,2020):
    # Read in each movie's URL from the previously scraped list of movies by year:
    df = pd.read_csv(str(year)+'_domestic_releases.csv')
    
    features=[]
    features_list=[]
    for i in range(0,len(df)):
        if i%20 == 0:
            print(year, i)
        moviepage=str(df.iloc[i][7])
        url='https://www.'+moviepage
        ##print(url)
        driver.get(url)
        html = driver.page_source
        soup = bs4(html, 'html.parser')

        list_text_rows=[]
        
        # Drill down to table area of relevance for capturing features:
        divs = soup.find_all('div',{'class':'mojo-summary-values'})
        for div in divs:
            rows = div.find_all('span')
            for row in rows:
                if(row.text.find('Domestic')):
                    list_text_rows.append(row.text)

        features = ['Budget','Genres','MPAA','Running Time']
        features_list.append(add_features_to_list(features, list_text_rows))
        
        #=================================
        #wait 1 seconds between iterations
        time.sleep(1)

    # Create feature columns and populate:
    temp = list(zip(*features_list))

    # Assign each column the feature values accumulated over all years:
    for i in range(0,3):
        df[features[i]] = list(temp[i])
    
    try:
        # Clean up Genres column
        df['Genres'] = df['Genres'].fillna('-')
        a = [g.rstrip() for g in df['Genres']]
        b = [re.sub('\n\s+',' ',s) for s in a]
        df['Genres'] = b
        
        
        # Write new dataframe to .csv file:
        df.to_csv(str(year)+'_movies_features_TEST.csv', index = False)
        
    except: # Skip a year if error encountered, re-run again separately if needed
        pass

# Kept for posterity: Finds a rarely seen error in 'Genres' element for movies on BOM. It has since been incorporated above.
##df['Genres'].isna().sum()
