==> 03_filter_movies_with_budget_info.ipynb <==
### Read in all 19,355 movies (1977-2019): clean and keep all movies (4341) that have budget info
***

# Import useful libraries:
import numpy as np
import pandas as pd

df_merged = pd.DataFrame()

for year in range(1977,2020):
    yr = str(year)
    df = pd.read_csv(yr+'_movies_features_TEST.csv')
    
    # rename erroneous 1st column:
    df.rename(columns={ df.columns[0]: 'Rank'}, inplace = True)

    # Clean up Budget column:
    df['Budget'] = df['Budget'].fillna('0')
    df['Budget'] = [int(s.replace('$','').replace(',','')) for s in df['Budget']]
    
    temp = df.take(df.loc[df['Budget'] > 0].index)
    #print('films with budget info,',yr,':',len(temp))
    
    df_merged = pd.concat([df_merged, temp.loc[:, temp.columns != 'Rank']], sort=False, ignore_index=True)

#print(len(df_merged))
#display(df_merged.tail())



# Add prefixes to URLs:
df_merged['URL'] = ['https://www.'+url for url in df_merged['URL']]

# write dataframe out to file for future use:
df_merged.to_csv('00_budgeted_movies_list.csv', index = False)
