==> 05_get_weekly_info_loopset1of4.ipynb <==
# Program to scrape the weekly info for the movie links supplied. Also scrapes a unique IMDB movie ID
import time
import regex as re
import numpy as np
import pandas as pd
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from bs4 import BeautifulSoup as bs4

chrome_options = Options()
chrome_options.add_argument('--disable-extensions')
chrome_options.add_argument('headless')
driver = webdriver.Chrome(executable_path='C:/bin/chromedriver.exe',options=chrome_options)

# Read in dataframe containing links:
#df_movies_list = pd.read_csv('00_budgeted_movies_list_TEST.csv')
df_movies_list = pd.read_csv('00_budgeted_movies_list.csv')

def get_weekly(text_rows):
    num_weeks = int(len(text_rows)/10)
    nmax = 15*10
    # if there is no weekly data, kick back a list of zeros to preserve the structure
    if num_weeks == 0:
        return [0]*15*5
    if num_weeks >= 15:
        a=text_rows[2:nmax:10]
        b=text_rows[3:nmax:10]
        c=text_rows[4:nmax:10]
        d=text_rows[5:nmax:10]
        e=text_rows[6:nmax:10]
        return a+b+c+d+e
    if num_weeks < 15: # 5 columns scraped over 15 rows (weeks), grouped into 1 output column
        return [0]*15*5
# ==============================
# ==============================
features_list = []
df_weekly_stats = pd.DataFrame()
features = ['Weekly','Theaters','Change','Avg','To Date']


#for i in range(0,len(df_movies_list)):
# files for loopset2of4, 3of4, and 4of4 do the exact same thing, just over different ranges of the total movies_list
for i in range(0,1000): # test range
    if i%20 == 0:
        print(i)
    try:
        movie_page=str(df_movies_list.loc[i]['URL'])
        url = movie_page+'weekly/'
        driver.get(url)
        html = driver.page_source
        soup = bs4(html, 'html.parser')
    except:
        print("1: An exception occurred")
    # ==============================
    # Get IMDB unique ID for subsequent scraping:
    #divs = soup.find_all('div',{'class':'a-box-inner'})
    divs = soup.select('div')
    
    pattern = 'title\/(.*?)\/'
    str_IMDB_ID = re.search(pattern,str(divs[0])).group(1)
    # ==============================
    # Get all weekly info for each movie:
    list_text_rows = []
    divs = soup.find_all('div',{'class':'mojo-gutter-sides'})
    for div in divs:
        rows = div.select('tr td')
        for row in rows:
            if(row.text.find('Date')):
                list_text_rows.append(row.text)

    
    df_weekly_stats[str_IMDB_ID] = get_weekly(list_text_rows)
    
    time.sleep(1)

# QC outputs: we want list_IMDB_IDs = 4341, and df_weekly_stats.shape = (75,4341)

# domain-decomposed: should be (75,1000)
print('corresponding weekly stats shape:',  df_weekly_stats.shape)

#### the shape of df_weekly_stats is not (75,4341) , meaning not all movie pages were successfully scraped

df_weekly_stats.head(10)

df_weekly_stats.to_csv('06_weekly_movies_features_part01.csv', index=True)

# Transpose the 'df_weekly_stats' dataframe so that each row's weekly stats correspond to a different movie:
df_new_features = df_weekly_stats.transpose()
print(df_weekly_stats.shape)
print(df_new_features.shape)
df_new_features.head(10)

df_new_features.to_csv('07_transposed_weekly_features_part01.csv', index=False)
