==> 06_get_IMDB_ratings_info_add_in_year_and_runtime.ipynb <==
# IMDB ratings file: title_ratings.tsv 
# (Publicly available at: https://www.imdb.com/interfaces/)
import pandas as pd

# Read in the IMDB ratings file:
df_imdb_basics = pd.read_csv('title_basics.tsv', sep='\t', dtype='str') # For release year and runtime in minutes

df_movie_list = pd.read_csv('01_budgeted_movie_list_with_IMDB_links.csv')

# Remove erroneous, non-numeric characters:
df_imdb_basics['startYear'] = df_imdb_basics['startYear'].str.extract('(\d+)', expand=False)
df_imdb_basics['runtimeMinutes'] = df_imdb_basics['runtimeMinutes'].str.extract('(\d+)', expand=False)

df_imdb_basics = df_imdb_basics.dropna(subset=['startYear'])
df_imdb_basics = df_imdb_basics.dropna(subset=['runtimeMinutes'])

# Re-cast into more appropriate data types:
df_imdb_basics['startYear']      = df_imdb_basics['startYear'].astype(int)
df_imdb_basics['runtimeMinutes'] = df_imdb_basics['runtimeMinutes'].astype(int)

print(len(df_imdb_basics))

df1 = df_movie_list['IMDB_ID']
df2 = df_imdb_basics['tconst']
df1=list(df1)
df2=list(df2)

#special case
df2.append('tt0046183')
df2.append('tt0029583')

# KEEP, Do not delete
# Get the indexes in df_imdb_ratings where they match the ones from df_movie_list:
res = [df2.index(i) for i in df1]
print(len(res))

year = [0.0]*len(df_movie_list)
mins = [0]*len(df_movie_list)

year = [x for ind, x in enumerate(df_imdb_basics['startYear'])]

df3 = df_imdb_basics['startYear']
df4 = df_imdb_basics['runtimeMinutes']

len(df3)

list_years = [df3[r] for r in res]
list_mins  = [df4[v] for v in res]

df_movie_list['Year'] = list_years
df_movie_list['Runtime Mins']  = list_mins

#df_movie_list.head()

df_movie_list.to_csv('02_budgeted_movies_list_with_ratings.csv', index=False)
