==> 07_concatenate_weekly_features_to_main.ipynb <==
import pandas as pd

df_movies = pd.read_csv('02_budgeted_movies_list_with_ratings.csv')
df_weekly_feat_01 = pd.read_csv('06_weekly_movies_features_part01.csv')
df_weekly_feat_02 = pd.read_csv('06_weekly_movies_features_part02.csv')
df_weekly_feat_03 = pd.read_csv('06_weekly_movies_features_part03.csv')
df_weekly_feat_04 = pd.read_csv('06_weekly_movies_features_part04.csv')

# Out of 4341 movies w/budget info, scraping shows only 793+806+836+1095=3530 have weekly info (seems off, will investigate).
print(df_movies.shape)
print(df_weekly_feat_01.shape)
print(df_weekly_feat_02.shape)
print(df_weekly_feat_03.shape)
print(df_weekly_feat_04.shape)

# 8 features identified from inspecting df_movies
#df_movies.head()

# 75 features shown (will definitely be whittled down) for each of the df_weekly_feat dataframes
a = df_weekly_feat_01.transpose()
b = df_weekly_feat_02.transpose()
c = df_weekly_feat_03.transpose()
d = df_weekly_feat_04.transpose()

# KEEP
# group transposed dataframes into a list, and then concat the list of dataframes
new_df  = pd.DataFrame()
list_of_transposed_df = [a,b,c,d]
new_df  = pd.concat(list_of_transposed_df)
new_df.head()

# Set the IMDB ID as the index for subsequent concat by index value.
temp = df_movies.copy()
temp.set_index('IMDB_ID',inplace=True)
display(temp.head())

# 'temp' is a proxy for 'df_movies'; 'new_df' is a proxy for the set of transposed weekly features 
#df_merged = temp.merge(new_df, left_index=True, right_index=True)
df_merged = temp.merge(new_df, left_index=True, right_index=True)
df_merged.shape

# drop duplicates
df_no_duplicates = df_merged.drop_duplicates(['Release','Release Date'],keep= 'last')
df_no_duplicates.shape

df_no_duplicates.to_csv('10_merged_movies_with_features_no_duplicates.csv')
