==> 08_NEW_data_cleansing_before_EDA.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime
from re import sub
from decimal import Decimal


***


#### Cleaning up raw dataframe: erroneous 1st column name, missing year in Release Date, replace NaNs

df_raw = pd.read_csv('10_merged_movies_with_features_no_duplicates.csv')

# Add year for Release Date column
#df_raw['Release Date'] = [i+' '+yr for i in df_raw['Release Date']]

# Replace NaNs
df_raw['Budget'] = df_raw['Budget'].fillna('0')
df_raw['MPAA']   = df_raw['MPAA'].fillna('0')

#df_raw.head()
df_raw.columns

# drop all rows if 1st weekly gross is not recorded:
print(df_raw.shape)
df_raw = df_raw[df_raw['0'] != '0']
print(df_raw.shape)
display(df_raw.head(10))


***


#### Assign movies a unique numerical ID

# SKIP
#movie_prefix = years[0]*10000
#
#df_new['MovieID'] = [movie_prefix+int(i) for i in df_raw[yr+' Rank']]
#print(df_new)

#df_raw.columns


***


#### Convert to numerical format: Gross box office values, # Theaters, Budget, and MPAA rating; fix datetime column

df_new = pd.DataFrame()
df_new[['Release','Release Date','Genres','URL']] = df_raw[['Release','Release Date','Genres','URL']]

df_new['Distributor'] = [d.replace('\n\n','') for d in df_raw['Distributor']]
#df_new['Distributor'] = df_raw['Distributor']

df_new['Calendar Gross'] = [Decimal(sub(r'[^\d.]', '', g)) for g in df_raw['Calendar Gross']]
df_new['Total Gross']    = [Decimal(sub(r'[^\d.]', '', g)) for g in df_raw['Total Gross']]
df_new['Max Theaters']   = [int(s.replace(',','')) for s in df_raw['Max Theaters']]

#df_new['Budget'] = [int(s.replace('$','').replace(',','')) for s in df_raw['Budget']]

temp_R = [rating.replace('NC-17','5').replace('R','4').replace('PG-13','3').replace('PG','2').replace('G','1') for rating in df_raw['MPAA']]
df_new['MPAA_numeric'] = [int(i) for i in temp_R]

##temp_DT = [datetime.strptime(date, '%b %d %Y') for date in df_raw['Release Date']]
##df_new['Release Date'] = [item.strftime('%d-%m-%Y') for item in temp_DT]

### Append unchanged columns to new dataframe:
##df_new[['Distributor','Genres']] = df_raw[['Distributor','Genres']]
##display(df_new)


***


# Insert the other feature columns:
df_new[['Rating','Votes','Budget']] = df_raw[['Rating','Votes','Budget']]
display(df_new.head())

# features: weekly gross (wkGROSS_)
for i in range(0,15):
    week_x_gross = 'wkGROSS_'+str(i+1)
    df_new[week_x_gross] = [Decimal(sub(r'[^\d.]', '', g)) for g in df_raw[str(i)]]
# ==============================

# features: weekly percent changes in gross (pctCHNG_)
##for i in range(15,30):
##    pct_chng = 'pctCHNG_'+str(i-14)
##    df_new[pct_chng] = [Decimal(sub(r'[^-\d.]', '', g.replace('-','0'))) for g in df_raw[str(i)]]
# ==============================


# features: current number of theatres showing the film (nTHEATERS_)
for i in range(30,45):
##for i in range(15,30):
    nTheaters = 'nTHEATERS_'+str(i-29)
    ##nTheaters = 'nTHEATERS_'+str(i-14)
    df_new[nTheaters]   = [int(s.replace(',','').replace('-','0')) for s in df_raw[str(i)]]
# ==============================


# features: weekly change in number of theaters (dTHEATERS_)
##for i in range(45,60):
##    dTheaters = 'dTHEATERS'+str(i-44)
##    df_new[dTheaters]   = [int(s.replace(',','')) for s in df_raw[str(i)]]
##    df_new[pTAG] = [Decimal(sub(r'[^\d.]', '', g.replace('-','0'))) for g in df_raw[str(i)]]
# ==============================

# features: Per Theater Average Gross (pTAG_)
for i in range(60,75):
##for i in range(30,45):
    pTAG = 'pTAG_'+str(i-59)
    ##pTAG = 'pTAG_'+str(i-29)
    df_new[pTAG] = [Decimal(sub(r'[^-\d.]', '', g.replace('-','0'))) for g in df_raw[str(i)]]

df_new.head()

# Write cleaned dataframe out to file:
df_new.to_csv('11_input_for_regression.csv')

##from datetime import datetime

#####df_raw['Release Date'] = [i+' '+yr for i in df_raw['Release Date']]

##a = [datetime.strptime(date, '%b %d %Y') for date in df_raw['Release Date']]
#print(date_obj.strftime('%d-%m-%Y'))

##a = [datetime.strptime(date, '%b %d %Y') for date in df_raw['Release Date']]
##b = [item.strftime('%d-%m-%Y') for item in a]
#print(a)
##print(b)

# Append unchanged columns to new dataframe:
##df_new[['Distributor','Genres']] = df_raw[['Distributor','Genres']]
##display(df_new)
